Added explicit API for Ohua.Types
Clamped down on `Ohua.Types` it now does not just export
everything. In fact most constructors are now hidden, to make a more
stable, API where we can change the underlying representation if we so
choose.
Access is handled via a `Make` and `Unwrap` typeclass to convert
between ohua types and standard types (safely).

Moved `HostExpr` to Ohua.Types (it makes more sense there)
Replaced `idName` in passes with the correct `Ohua.ALang.Refs.id`
Adapted other modules to use the `Make` and `Unwrap` class as part of
the new `Types` API
Added a module for special HostExpr constants
Encountered an aeson error for which there is now an example in `misc`